Issue with koin injection for the alarm scheduler
