+++ composeApp/src/androidMain/kotlin/org/example/project/androidSpecificModule.kt
@@ -1,7 +1,6 @@
 package org.example.project
 
 import org.example.project.alarm.data.AlarmScheduler
-import org.example.project.alarm.data.AndroidAlarmSchedulerImp
 import org.koin.android.ext.koin.androidApplication
 import org.koin.dsl.module
